Added working Jacoco plugin

+++ src/main/java/com/maveric/accountservice/controller/AccountController.java
@@ -5,6 +5,7 @@ import com.maveric.accountservice.dto.BalanceDto;
 import com.maveric.accountservice.dto.TransactionDto;
 import com.maveric.accountservice.dto.UserDto;
 import com.maveric.accountservice.exception.CustomerNotFoundException;
+import com.maveric.accountservice.exception.UnAuthorizedException;
 import com.maveric.accountservice.feignconsumer.BalanceServiceConsumer;
 import com.maveric.accountservice.feignconsumer.TransactionServiceConsumer;
 import com.maveric.accountservice.feignconsumer.UserServiceConsumer;
@@ -15,8 +16,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
+
+import static com.maveric.accountservice.constants.Constants.AUTH_HEADER_ERROR_MESSAGE;
 
 
 @RestController
@@ -33,7 +37,7 @@ public class AccountController {
     @Autowired
     UserServiceConsumer userServiceConsumer;
 
-    /* Returns list of all accounts */
+    /* Dummy Returns list of all accounts */
     @GetMapping("customers/{customerId}/account")
     public ResponseEntity<List<AccountDto>> getAccounts(@PathVariable String customerId,@RequestParam(defaultValue = "0") Integer page,
                                                         @RequestParam(defaultValue = "10") Integer pageSize) {
@@ -44,26 +48,37 @@ public class AccountController {
     /* Returns list of all valid accounts belonging a particular Customer Id */
     @GetMapping("customers/{customerId}/accounts")
     public ResponseEntity<List<AccountDto>> getAccountByCustomerId(@PathVariable String customerId,@RequestParam(defaultValue = "0") Integer page,
-                                                                   @RequestParam(defaultValue = "5") Integer pageSize) {
-        List<AccountDto> accountDtoResponse = accountService.getAccountByUserId(page,pageSize,customerId);
-        return new ResponseEntity<>(accountDtoResponse, HttpStatus.OK);
-    }
+                                                                   @RequestParam(defaultValue = "5") Integer pageSize,@RequestHeader String userEmail) {
+        ResponseEntity<UserDto> responseEntityUserDto = userServiceConsumer.getUserDetails(customerId);
+             UserDto userDto = responseEntityUserDto.getBody()==null?new UserDto():responseEntityUserDto.getBody();
+            if (userDto.get_id() != null) {
+                if (userDto.getEmail().equalsIgnoreCase(userEmail)) {
+                    List<AccountDto> accountDtoResponse = accountService.getAccountByUserId(page, pageSize, customerId);
+                    return new ResponseEntity<>(accountDtoResponse, HttpStatus.OK);
+                } else {
+                    throw new UnAuthorizedException(AUTH_HEADER_ERROR_MESSAGE);
+                }
+            }
+            else {
+                throw new CustomerNotFoundException("Account details not available because Customer Id-" + customerId + " not found.");
+            }
+        }
+
 
     /* Creates a valid account */
     @PostMapping("customers/{customerId}/accounts")
-    public ResponseEntity<AccountDto> createAccount(@PathVariable String customerId, @Valid @RequestBody AccountDto accountDto) {
-        UserDto userDto = null;
-        try {
+    public ResponseEntity<AccountDto> createAccount(@PathVariable String customerId, @Valid @RequestBody AccountDto accountDto,@RequestHeader String userEmail) {
             ResponseEntity<UserDto> responseEntityUserDto = userServiceConsumer.getUserDetails(accountDto.getCustomerId());
-            userDto = responseEntityUserDto.getBody();
-        }
-        catch(Exception ex)
-        {
-            ex.printStackTrace();
-        }
-        if(userDto!=null) {
-            AccountDto accountDtoResponse = accountService.createAccount(customerId, accountDto);
-            return new ResponseEntity<>(accountDtoResponse, HttpStatus.CREATED);
+        @NotNull UserDto userDto = responseEntityUserDto.getBody();
+
+        if(userDto.get_id()!=null) {
+            if(userDto.getEmail().equalsIgnoreCase(userEmail)) {
+                AccountDto accountDtoResponse = accountService.createAccount(customerId, accountDto);
+                return new ResponseEntity<>(accountDtoResponse, HttpStatus.CREATED);
+            }
+            else {
+                throw new UnAuthorizedException(AUTH_HEADER_ERROR_MESSAGE);
+            }
         }
         else {
             throw new CustomerNotFoundException("Customer Id-"+accountDto.getCustomerId()+" not found. Cannot create account.");
@@ -72,47 +87,55 @@ public class AccountController {
 
     /* Returns Account details along with Balance and Transaction details based on valid Account Id */
     @GetMapping("customers/{customerId}/accounts/{accountId}")
-    public ResponseEntity<AccountDto> getAccountDetails(@PathVariable String customerId,@PathVariable String accountId) {
-        AccountDto accountDtoResponse = accountService.getAccountDetailsById(accountId);
+    public ResponseEntity<AccountDto> getAccountDetails(@PathVariable String customerId,@PathVariable String accountId,@RequestHeader String userEmail) {
 
-        try {
-            ResponseEntity<BalanceDto> balanceDto = balanceServiceConsumer.getBalances(accountId);
-            accountDtoResponse.setBalance(balanceDto.getBody());
+        ResponseEntity<UserDto> responseEntityUserDto = userServiceConsumer.getUserDetails(customerId);
+        UserDto userDto = responseEntityUserDto.getBody()==null?new UserDto():responseEntityUserDto.getBody();
+        if(userDto.getEmail().equalsIgnoreCase(userEmail)) {
+            AccountDto accountDtoResponse = accountService.getAccountDetailsById(accountId);
+                ResponseEntity<BalanceDto> balanceDto = balanceServiceConsumer.getBalances(accountId);
+                accountDtoResponse.setBalance(balanceDto.getBody());
+
+            return new ResponseEntity<>(accountDtoResponse, HttpStatus.OK);
         }
-        catch(Exception ex)
+        else
         {
-            accountDtoResponse.setBalance(new BalanceDto());
+            throw new UnAuthorizedException(AUTH_HEADER_ERROR_MESSAGE);
         }
-
-
-        try {
-            ResponseEntity<List<TransactionDto>> transactionDto = transactionServiceConsumer.getTransactionsByAccountId(accountId);
-            accountDtoResponse.setTransactions(transactionDto.getBody());
-        }
-        catch(Exception ex)
-        {
-            accountDtoResponse.setTransactions(new ArrayList<>());
-        }
-
-        return new ResponseEntity<>(accountDtoResponse, HttpStatus.OK);
     }
 
     /* Updates account details based on valid Account Id */
     @PutMapping("customers/{customerId}/accounts/{accountId}")
-    public ResponseEntity<AccountDto> updateAccount(@PathVariable String customerId,@PathVariable String accountId,@RequestBody AccountDto accountDto) {
-        AccountDto accountDtoResponse = accountService.updateAccountDetails(customerId,accountId,accountDto);
-        return new ResponseEntity<>(accountDtoResponse, HttpStatus.OK);
+    public ResponseEntity<AccountDto> updateAccount(@PathVariable String customerId,@PathVariable String accountId,@RequestBody AccountDto accountDto,@RequestHeader String userEmail) {
+        ResponseEntity<UserDto> responseEntityUserDto = userServiceConsumer.getUserDetails(customerId);
+        UserDto userDto = responseEntityUserDto.getBody()==null?new UserDto():responseEntityUserDto.getBody();
+        if(userDto.getEmail().equalsIgnoreCase(userEmail)) {
+            AccountDto accountDtoResponse = accountService.updateAccountDetails(customerId, accountId, accountDto);
+            return new ResponseEntity<>(accountDtoResponse, HttpStatus.OK);
+        }
+        else
+        {
+            throw new UnAuthorizedException(AUTH_HEADER_ERROR_MESSAGE);
+        }
+
     }
 
     /* Deletes account along with its corresponding balance and transaction details based on valid Account Id */
     @DeleteMapping("customers/{customerId}/accounts/{accountId}")
-    public ResponseEntity<String> deleteAccount(@PathVariable String customerId,@PathVariable String accountId) {
-        String result = accountService.deleteAccount(accountId);
-        if(result!=null)
-        {
-            balanceServiceConsumer.deleteBalanceByAccountId(accountId);
-            transactionServiceConsumer.deleteTransactionByAccountId(accountId);
+    public ResponseEntity<String> deleteAccount(@PathVariable String customerId,@PathVariable String accountId,@RequestHeader String userEmail) {
+        ResponseEntity<UserDto> responseEntityUserDto = userServiceConsumer.getUserDetails(customerId);
+        UserDto userDto = responseEntityUserDto.getBody()==null?new UserDto():responseEntityUserDto.getBody();
+        if(userDto.getEmail().equalsIgnoreCase(userEmail)) {
+            String result = accountService.deleteAccount(accountId);
+            if (result != null) {
+                balanceServiceConsumer.deleteBalanceByAccountId(accountId);
+                transactionServiceConsumer.deleteTransactionByAccountId(accountId);
+            }
+            return new ResponseEntity<>(result, HttpStatus.OK);
         }
-        return new ResponseEntity<>(result, HttpStatus.OK);
+        else
+        {
+            throw new UnAuthorizedException(AUTH_HEADER_ERROR_MESSAGE);
+        }
     }
 }
